Move post_mixer to Views module
