+++ apps/main/lib/main/views/post_mixer.rb
@@ -1,5 +1,5 @@
 module Main
-  module Persistence
+  module Views
     class PostMixer
 
       attr_reader :posts
